endpoint: ensure we have a hid device ready from the kernel
As stated by `create_kernel_device()` in uhid.py:
```
Note that the device is not immediately ready to go after creation,
you must wait for :meth:`start` and ideally for :meth:`open` to be
called.
```

We can also simply wait for the udev_device to be set, because that's
the only thing we need to care here.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>