tests: simplify catch_evdev_events
Instead of setting up the threads whenever we need to listen to evdev
events, we might as well setting it up once during the creation of
the Device. Each Endpoint gets notified when a new child device gets
added to its sysfs tree, so we can fetch that information to open the
corresponding nodes at that time.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>